+++ src-docs/src/views/side_nav/side_nav.js
@@ -4,8 +4,6 @@ import React, {
 
 import {
   EuiSideNav,
-  EuiSideNavItem,
-  EuiSideNavTitle,
 } from '../../../../src/components';
 
 export default class extends Component {
@@ -24,42 +22,37 @@ export default class extends Component {
   };
 
   render() {
+    const sideNav = [{
+      name: 'Kibana',
+      id: 0,
+      items: [{
+        name: 'Advanced settings',
+        id: 1,
+        onClick: () => { window.alert('Advanced settings'); },
+      }, {
+        name: 'Index Patterns (link)',
+        id: 2,
+        href: 'http://www.elastic.co',
+      }, {
+        name: 'Saved Objects',
+        id: 3,
+        onClick: () => { window.alert('Saved Objects'); },
+        isSelected: true,
+      }, {
+        name: 'Reporting',
+        id: 4,
+        onClick: () => { window.alert('Reporting'); },
+      }],
+    }];
+
     return (
       <EuiSideNav
         mobileTitle="Navigate within $APP_NAME"
         toggleOpenOnMobile={this.toggleOpenOnMobile}
         isOpenOnMobile={this.state.isSideNavOpenOnMobile}
         style={{ width: 192 }}
-      >
-
-        <EuiSideNavTitle>
-          Kibana
-        </EuiSideNavTitle>
-        <EuiSideNavItem>
-          <button>
-            Advanced settings
-          </button>
-        </EuiSideNavItem>
-
-        <EuiSideNavItem isSelected>
-          <button>
-            Index Patterns
-          </button>
-        </EuiSideNavItem>
-
-        <EuiSideNavItem>
-          <button>
-            Saved Objects
-          </button>
-        </EuiSideNavItem>
-
-        <EuiSideNavItem>
-          <button>
-            Reporting
-          </button>
-        </EuiSideNavItem>
-
-      </EuiSideNav>
+        items={sideNav}
+      />
     );
   }
 }