+++ src/components/index.js
@@ -188,9 +188,6 @@ export {
 
 export {
   EuiSideNav,
-  EuiSideNavGroup,
-  EuiSideNavItem,
-  EuiSideNavTitle,
 } from './side_nav';
 
 export {