+++ src/components/side_nav/side_nav.test.js
@@ -33,4 +33,92 @@ describe('EuiSideNav', () => {
         .toMatchSnapshot();
     });
   });
+
+  describe('items', () => {
+    test('is rendered', () => {
+      const sideNav = [{
+        name: 'A',
+        id: 0,
+        items: [{
+          name: 'B',
+          id: 1,
+        }, {
+          name: 'C',
+          id: 2,
+          items: [{
+            name: 'D',
+            id: 3,
+          }, {
+            name: 'E',
+            id: 4,
+          }],
+        }],
+      }];
+
+      const component = render(
+        <EuiSideNav items={sideNav} />
+      );
+
+      expect(component)
+        .toMatchSnapshot();
+    });
+
+    test('renders items with are links', () => {
+      const sideNav = [{
+        name: 'A',
+        id: 0,
+        href: 'http://www.elastic.co',
+        items: [{
+          name: 'B',
+          id: 1,
+        }, {
+          name: 'C',
+          id: 2,
+          items: [{
+            name: 'D',
+            id: 3,
+          }, {
+            name: 'E',
+            id: 4,
+          }],
+        }],
+      }];
+
+      const component = render(
+        <EuiSideNav items={sideNav} />
+      );
+
+      expect(component)
+        .toMatchSnapshot();
+    });
+
+    test('renders selected item and automatically opens parent items', () => {
+      const sideNav = [{
+        name: 'A',
+        id: 0,
+        items: [{
+          name: 'B',
+          id: 1,
+        }, {
+          name: 'C',
+          id: 2,
+          items: [{
+            name: 'D',
+            id: 3,
+            isSelected: true,
+          }, {
+            name: 'E',
+            id: 4,
+          }],
+        }],
+      }];
+
+      const component = render(
+        <EuiSideNav items={sideNav} />
+      );
+
+      expect(component)
+        .toMatchSnapshot();
+    });
+  });
 });